cambio de estilos

+++ src/app/components/panel/panel.component.ts
@@ -1,11 +1,14 @@
 import { Component, OnInit, NgZone, ComponentFactoryResolver, OnDestroy, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 
+import { DialogService } from 'src/app/shared/dialog/dialog.service';
+
 import { LoginService } from '../login/login.service';
 
 import { SectionDirective } from './section.directive';
 import { SectionItem } from './section.item';
 import { SectionService } from './section.service';
+import { Dialog } from 'electron';
 
 @Component({
   selector: 'app-panel',
@@ -16,13 +19,15 @@ export class PanelComponent implements OnInit, OnDestroy {
   @ViewChild(SectionDirective, {static:true}) sectionHost: SectionDirective;
 
   sections: SectionItem[];
+  sectionSelected = 0;
 
   constructor(
     private _ngZone: NgZone,
     private _router: Router,
     private _cfr: ComponentFactoryResolver,
     private _loginService: LoginService,
-    private _sectionService: SectionService
+    private _sectionService: SectionService,
+    private _dialogService: DialogService
   ) { }
 
   ngOnInit(): void {
@@ -36,16 +41,18 @@ export class PanelComponent implements OnInit, OnDestroy {
 
   setInfoComponent() {
     this.loadComponent(this.sections[0]);
+    this.sectionSelected = 0;
   }
 
   setEventosComponent() {
     this.loadComponent(this.sections[1]);
+    this.sectionSelected = 1;
   }
 
   setGaleriaComponent() {
     this.loadComponent(this.sections[2]);
+    this.sectionSelected = 2;
   }
-
 
   loadComponent(sectionItem: SectionItem) {
     const componentFactory = this._cfr.resolveComponentFactory(sectionItem.component);
@@ -55,10 +62,16 @@ export class PanelComponent implements OnInit, OnDestroy {
   }
 
   logout() {
-    this._loginService.logout().then(
-      () => {
-        this._ngZone.run(() => this._router.navigate(['']));
-        localStorage.clear();
+    this._dialogService.buildConfirmationDialog('Salir','¿Desea cerrar sesión?').subscribe(
+      result => {
+        if(result) {
+          this._loginService.logout().then(
+            () => {
+              this._ngZone.run(() => this._router.navigate(['']));
+              localStorage.clear();
+            }
+          );
+        }
       }
     );
   }
